refactor the orderTasks function
if run, this commit refactor the orderTasks function to get the task that will be realocated and the task that will give the position for her.  the logic is: delete from the list tasks the task to be relocated, get index of task that will give the position and make a insert using the created index and the task.