altering some conditionals in use_case and finalizing the tests
i made a mistake making the conditional "if" validate by index if the value exists in the list, so i changed to see if the task is in the tasklists.

If applied, this commit will change what was explained above and call the use cases to tests.py to finalizing the tests functions.