creating a treatment for improper type of parameter and testing it
if run, this commit will create a if stantment to check that the type of parameters are correct. If not, i create a message error and call the custom error class.

The getters and setters are working only if the user call the print of todolist, but i want to check before the object creation if all parameters is ok.