creating some operations in the repositories
if run, this commit will create, in todo_list_repository, some functions to make operations in database. the function "get_the_next_free_position" takes the position of the last task. this function will be used to generate a new position when the user is going to create a new task.
Also, in the user_repository, i create a task to verify if user email is registered.