Creating a test to insert a task in database and fix bugs in get_next_free_position
If run, this commit will create a test to insert a task in a list. For this test, i need to extract the  list id in which the user wants to place the object and extract the next free position in this same list, to place the task. When the use cases are implemented, i wanna use a synchronized function to make ensure no one insert a task while i don't extract the next free position.