[1/2] Refactor a operation of creation in TodoList package, creating another function in repository and start the alterations about some tests.
If run, this commit will be refactor the create_task function to receives one more parameter, a todolist_id. I made this change because, in my database, my Task need a todolist_id to be created. taking this in consideration, i decided to create a function in my repository to recover a the id of todolist using the name of the list and the email of creator. Making this, i can leave the database create the ID and i can extract the id using another parameters.
But it make my system not accept a user create two or more lists with the same name or two or more tasks with the same parameters.