Applying the alterations made in the last commit
This commit is large, but is the same change for all the functions. How i create a function to recover my list id by the name and email and i increase a parameter "email" in todoList constructor, i need to pass a email now, and using the todolist datas to recover your id from the database.