removing a task list from todolist.
if run, this commit wll remove the attribute "tasks" from todolist package. I made this because now the objects are stored in database. Into the tbTask, i pass the Todolist Id like a Foreign key, to relate the task to the todolist. Using this method, the attribute "tasks" it is no longer necessary to relate tasks to a list.